requirements.txt
beautifulsoup4==4.14.2
Flask==3.1.2
flask-cors==6.0.1
langchain==1.0.8
langchain-community==0.4.1
langchain-openai==1.0.3
lxml==6.0.2
python>=3.10.12
pydantic==2.12.4
pytest==9.0.1
python-dotenv==1.2.1
requests==2.32.5